+++ NEWS.md
@@ -4,6 +4,7 @@
 * Fixed the `poa_nsw` dataset to avoid strange geometries near the state border.
 * Added geometries for Lord Howe Island, Australian Capital Territory, and
   Jervis Bay Territory, and a utility function `outline()` for using them.
+* Add Greater Sydney boundary as `sydney` dataset.
 * Slightly increase the resolution to 750 m instead of 1 km.
 * Update to the 2023 ASGS release, including the 2023 LGA shapefile.